requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-rest-knox==4.2.0
PyJWT==2.8.0
cryptography==41.0.7
pyotp==2.9.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Email
django-anymail==10.2

# File Upload & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.10

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.0

# Utilities
setuptools==80.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
ldap3==2.9.1

# Monitoring & Logging
# Optional error reporting
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3

# Performance
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0

# Search
elasticsearch==8.11.1
django-elasticsearch-dsl==8.0

# Notifications
twilio==8.11.1
sendgrid==6.11.0

# Excel/CSV Export
openpyxl==3.1.2
pandas==2.1.4

# Validation
marshmallow==3.20.1
